Rename files for easier fuzzy finding
