+++ lib/ynab_convert/documents.rb
@@ -2,11 +2,15 @@
 
 # Groups Statements and YNAB4File
 module Documents
-  documents = %w[statements ynab4_files]
+  documents = %w[statement ynab4_file]
 
   # Load all known Documents
-  documents.each do |document|
-    Dir[File.join(__dir__, 'documents', document, '*.rb')].each do |file|
+  documents.each do |d|
+    # Require the base classes first so that its children can find the parent
+    # class since files are otherwise loaded in alphabetical order
+    require File.join(__dir__, 'documents', "#{d}s", "#{d}.rb")
+
+    Dir[File.join(__dir__, 'documents', "#{d}s", '*.rb')].each do |file|
       require file
     end
   end
