+++ lib/ynab_convert/transformers/formatters/n26_formatter.rb
