+++ lib/ynab_convert/validators/ynab4_row_validator.rb
