+++ spec/documents/statements/n26_statement_spec.rb
@@ -1,13 +1,12 @@
 # frozen_string_literal: true
 
 require 'ynab_convert/documents/statements/statement'
-require 'ynab_convert/documents/statements/n26'
+require 'ynab_convert/documents/statements/n26_statement'
 
 RSpec.describe Documents::Statements::N26 do
   statement = File.join(File.dirname(__dir__),
                         '..',
-                        'fixtures/documents/statements',
-                        'n26/n26.csv')
+                        'fixtures/statements/n26.csv')
 
   let(:subject) { Documents::Statements::N26.new(filepath: statement) }
