+++ spec/transformers/cleaners/n26_cleaner_spec.rb
@@ -1,7 +1,7 @@
 # frozen_string_literal: true
 
 require 'ynab_convert/transformers/cleaners/cleaner'
-require 'ynab_convert/transformers/cleaners/n26'
+require 'ynab_convert/transformers/cleaners/n26_cleaner'
 
 RSpec.describe Transformers::Cleaners::N26 do
   let(:statement) do
