+++ spec/transformers/enhancers/n26_enhancer_spec.rb
@@ -1,6 +1,6 @@
 # frozen_string_literal: true
 
-require 'ynab_convert/transformers/enhancers/n26'
+require 'ynab_convert/transformers/enhancers/n26_enhancer'
 
 RSpec.describe Transformers::Enhancers::N26 do
   let(:subject) { Transformers::Enhancers::N26.new }
